build databased for NPN4 xags
